fix(users): move fetch users to utils folder
