--- galasa-ui/src/utils/uesrs.ts
@@ -1,5 +1,0 @@
-/*
- * Copyright contributors to the Galasa project
- *
- * SPDX-License-Identifier: EPL-2.0
- */
